Render update
Got rid of deprecated SDL_BlitSurface (SDL1) and replaced it with
SDL_RenderCopy. In the words of the creators of SDL:

/* The renderer works pretty much like a big canvas:
when you RenderCopy you are adding paint, each time adding it
on top.
You can change how it blends with the stuff that
the new data goes over.
When your 'picture' is complete, you show it
by using SDL_RenderPresent. */

/* SDL 1.2 hint: If you're stuck on the whole renderer idea coming
from 1.2 surfaces and blitting, think of the renderer as your
main surface, and SDL_RenderCopy as the blit function to that main
surface, with SDL_RenderPresent as the old SDL_Flip function.*/