Fixed Previous fucked Commits. READ COMMENTS
Have ensured map1.bin will no longer sync, as it is now 400mb or so. You
will have to re-generate it on your own. CInGameState contains an IF
statement for this. It is commented out ATM. Run the program once, then
uncomment it so it is back in. This will prevent you from creating a
100,000,000 point heightmap every time you load.
Otherwise, terrain streaming is done, but lags like a bitch during the
load function. I'm working on it, I'll have a fix soon.